pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "xxcli"
version = "0.2.0"
description = "A CLI for consuming and contributing to Twitter/X without opening the browser."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{ name = "Declan Kramper", email = "declan@dkbuilds.co" }]
keywords = ["twitter", "x", "cli", "social-media"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Utilities",
]
dependencies = [
    "tweepy>=4.14",
    "click>=8.0",
    "rich>=13.0",
    "openai-agents>=0.1",
    "pyyaml>=6.0",
]

[project.scripts]
xx = "xxcli.cli:main"

[project.urls]
Homepage = "https://github.com/declankra/xxcli"
Repository = "https://github.com/declankra/xxcli"
Issues = "https://github.com/declankra/xxcli/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-ra"
